+++ gradle.properties
@@ -12,7 +12,7 @@ loader_version=0.16.9
 
 
 # Mod Properties
-	mod_version = 0.3-beta+mc1.21.3
+	mod_version = 0.3.0-beta+mc1.21.3
 	maven_group = xyz.telosaddon.yuno
 	archives_base_name = realmsaddon
