+++ src/main/java/xyz/telosaddon/yuno/ui/AbstractCustomElement.java
@@ -4,7 +4,7 @@ import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.DrawContext;
 import xyz.telosaddon.yuno.TelosAddon;
 import xyz.telosaddon.yuno.sound.SoundManager;
-import xyz.telosaddon.yuno.utils.config.TelosConfig;
+
 
 public abstract class AbstractCustomElement implements CustomElement{
