+++ src/main/resources/assets/telosaddon/lang/en_us.json
@@ -2,5 +2,6 @@
   "category.telosaddon" : "Telos Addon",
   "key.telosaddon.call": "Call Hotkey",
   "key.telosaddon.menu": "Menu Hotkey",
-  "key.telosaddon.nexus": "Nexus Hotkey"
+  "key.telosaddon.nexus": "Nexus Hotkey",
+  "key.telosaddon.tpmenu": "TP Menu Hotkey"
 }
