+++ src/main/resources/assets/telosaddon/owo_ui/rangetab.xml
@@ -3,6 +3,9 @@
     <components>
         <flow-layout direction="vertical">
             <children>
+                <label>
+                    <text>RealmsAddon v0.3.0</text>
+                </label>
                 <flow-layout direction="horizontal">
                     <children>
                         <button id="Home">
@@ -27,16 +30,6 @@
                         </button>
                         <button id="Settings">
                             <text>Settings</text>
-                            <margins>
-                                <left>5</left>
-                                <top>5</top>
-                            </margins>
-                            <sizing>
-                                <horizontal method="fixed">100</horizontal>
-                            </sizing>
-                        </button>
-                        <button id="Teleport">
-                            <text>Teleport</text>
                             <margins>
                                 <left>5</left>
                                 <top>5</top>
